# openapi.yml
openapi: 3.0.3
info:
  title: Resourcery API
  description: API documentation for the Resourcery community resource booking system.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Spring Boot — Auth Service
  - url: http://localhost:8081
    description: PHP — Resources Service
  - url: http://localhost:8082
    description: Django — Bookings & Notifications Service
...
paths:
...
  /bookings:
    post:
      tags: [Bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingInput'
      responses:
        '201':
          description: Booking created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '200':
          description: Existing pending booking updated (merged)
        '409':
          description: Conflict — dates taken or extension prompt
...
  /bookings/{id}/status:
    patch:
      tags: [Bookings]
      summary: Update booking status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [for_pickup, declined, not_returned, returned, pending]
      responses:
        '200':
          description: Status updated
        '409':
          description: Conflict — dates already booked by another user
  /bookings/{id}/extend:
    patch:
      tags: [Bookings]
      summary: Extend or update an approved/not-returned booking
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [merged_start, merged_end]
              properties:
                merged_start: { type: string, format: date }
                merged_end: { type: string, format: date }
                is_extension: { type: boolean }
      responses:
        '200':
          description: Booking updated, reset to pending
...
  /bookings/user/{username}:
    delete:
      tags: [Bookings]
      summary: Delete all active bookings for a member (admin only)
      parameters:
        - name: username
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Bookings removed

# ── NOTIFICATION ENDPOINTS (Django) ──────────────────────────────────────────
...
components:
  schemas:
...
    Booking:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        resource_id: { type: integer }
        resource_name: { type: string }
        resource_type: { type: string }
        date: { type: string, format: date }
        return_date: { type: string, format: date, nullable: true }
        time: { type: string, format: time }
        end_time: { type: string, format: time, nullable: true }
        quantity_requested: { type: integer, default: 1 }
        status:
          type: string
          enum: [pending, for_pickup, not_returned, returned, declined, cancelled, no_pickup]
        is_extension: { type: boolean }
        is_update: { type: boolean }
        created_at: { type: string, format: date-time }
    BookingInput:
      type: object
      required: [resource_id, resource_name, date, time]
      properties:
        resource_id: { type: integer }
        resource_name: { type: string }
        resource_type: { type: string }
        date: { type: string, format: date }
        return_date: { type: string, format: date, nullable: true }
        time: { type: string, format: time }
        end_time: { type: string, format: time, nullable: true }
        quantity_requested: { type: integer, default: 1 }
...